Prevent package-lock.json creation
Since the package-lock.json is never part of the published package, having it in the repo makes the CI tests to be run with different versions of the dependencies than the ones a fresh install by a user would fetch. As a result the CI will never surface regressions that might have been caused in the latest version of our dependencies and we will not realize that fresh install might have broke b/c of that.